tool/requirements-telegram.txt
# Local-only Telegram insights API (run: ./run_telegram_api.sh)
telethon>=1.36.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
# Optional: image understanding + sharper meme lines (server .env only)
openai>=1.40.0
